+++ components/images/image-detail/ImageDetailModal.tsx
@@ -118,7 +118,7 @@ export default function ImageDetailModal({
             height={400}
           />
         ) : (
-          // eslint-disable-next-line @next/next/no-img-element
+           
           <img
             src={imageSrc}
             alt={image.filename}